requirements.txt
Flask==2.3.2
pandas==2.2.2
gspread==6.0.2
oauth2client==4.1.3
python-dateutil==2.8.2
